Move tests into packing/fields test fields
